Skills/설문 연동/requirements-forms.txt
# Google Forms API (question_to_google_form.py)
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
